Refactor ExWire.Adapter.TCP
Why does it need a refactor?
===========================

This GenServer is managing the TCP connection, the aggregation of TCP
data into wire protocol "packets" (e.g. GetBlockHeaders packet), as well
as initializing the session of the Devp2p Wire Protocol (which includes
crypto and protocol handshakes, and the sending and receiving an initial
Hello message).

It should not manage so many things. It is currently complicated, and
GenServers should be as simple as possible, delegating most of the other
word to modules.

What was done?
=============

This refactor attempts to simplify what TCP is doing. Work done includes,

- Removing the notion of an "active" state. That is the session
responsibility and should live under some DEVp2p module.

- Remove any notion of the `Hello` message. That message is part of the
protocol handshake, and TCP need not concern itself with it.

- Clean up framing and unframing. We decided to leave those in this
GenServer for now. It seems like a good place to handle framing since we
can make it transparent for those using the TCP gen server.

- Remove responsibility for handling packets. Once the gen server has
handled the crypto handshake (sending auth and receiving ack), it should
simply forward any packets to subscribers. It should not handle the
packets themselves. That is what this module now does.

- Create some convenience client functions