First pass at DEVp2p Wire Protocol
First pass at handling the DEVp2p Wire Protocol. We have a session that
stores the packets sent and received for the handshake. We also check
for the `cap`s for compatibility. The session is only active if the
handshake is complete and if they are compatible.

We also handle `Disconnect` and `Ping` messages by responding
appropriately.